refactoring split in subpackages
