+++ celements-spring-security/src/main/java/com/celements/spring/security/api/IdentityService.java
@@ -1,4 +1,4 @@
-package com.celements.spring.security;
+package com.celements.spring.security.api;
 
 import javax.validation.constraints.NotEmpty;
 import javax.validation.constraints.NotNull;
