+++ celements-spring-security/src/main/java/com/celements/spring/security/auth/WikiClientRegistrationRepository.java
@@ -1,4 +1,4 @@
-package com.celements.spring.security;
+package com.celements.spring.security.auth;
 
 import javax.annotation.Nullable;
 import javax.inject.Inject;
@@ -11,6 +11,8 @@ import org.springframework.security.oauth2.client.registration.ClientRegistratio
 import org.springframework.security.oauth2.core.AuthorizationGrantType;
 import org.springframework.security.oauth2.core.oidc.IdTokenClaimNames;
 import org.springframework.stereotype.Component;
+
+import com.celements.spring.security.api.IdentityService;
 
 @Component
 public class WikiClientRegistrationRepository implements ClientRegistrationRepository {
