+++ celements-spring-security/src/main/java/com/celements/spring/security/web/CookieBearerTokenResolver.java
@@ -1,4 +1,4 @@
-package com.celements.spring.security;
+package com.celements.spring.security.web;
 
 import javax.servlet.http.HttpServletRequest;
